PF-48-просмотр-своих-постов : fix багов, вынес типизацию отдельно. Сделал Универсальный компонент для условного рендера блока навигации для гостя, своей страницы и не авторизованного пользователя
