+++ src/lib/feature/profile/api/profileApi.ts
@@ -1,62 +1,9 @@
 import { baseApi } from '@/lib/baseApi'
-import { UserProfileDataResponse } from '@/app/profile/[id]/page'
-
-type AvatarType = {
-  url: string
-  width: number
-  height: number
-  fileSize: number
-  createdAt: string
-}
-
-type ProfileType = {
-  id: number
-  userName: string
-  firstName: string
-  lastName: string
-  city: string
-  country: string
-  region: string
-  dateOfBirth: string
-  aboutMe: string
-  createdAt: string
-  avatars: AvatarType[]
-}
-
-type UserPostsResponse = {
-  totalCount: number
-  pageSize: number
-  items: [
-    {
-      id: number
-      userName: string
-      description: string
-      location: string
-      images: [
-        {
-          url: string
-          width: number
-          height: number
-          fileSize: number
-          createdAt: string
-          uploadId: string
-        },
-      ]
-      createdAt: string
-      updatedAt: string
-      avatarOwner: string
-      ownerId: number
-      owner: {
-        firstName: string
-        lastName: string
-      }
-      likesCount: number
-      isLiked: boolean
-      avatarWhoLikes: []
-    },
-  ]
-  totalUsers: number
-}
+import {
+  ProfileType,
+  UserPostsResponse,
+  UserProfileDataResponse,
+} from '@/lib/feature/profile/types/profile.types'
 
 export const profileApi = baseApi.injectEndpoints({
   endpoints: build => ({
@@ -72,7 +19,7 @@ export const profileApi = baseApi.injectEndpoints({
     >({
       query: ({ userId, endCursor, pageSize }) =>
         `/posts/user/${userId}/${endCursor}?pageSize=${pageSize}`,
-      serializeQueryArgs: ({ endpointName }) => endpointName,
+      serializeQueryArgs: ({ endpointName, queryArgs }) => endpointName + queryArgs.userId,
       merge: (currentCache, newData) => {
         currentCache.items.push(...newData.items)
       },
