+++ src/lib/feature/profile/ui/components/profile/userMetadata/UserMetadata.tsx
@@ -4,17 +4,17 @@ import { Typography } from '@/components/ui/typography/Typography'
 
 type Props = {
   title: string
-  followers: number
+  metaDataCount: number
   // variant?: 'followers' | 'following'
 }
 
-export const UserMetadata = ({ title, followers }: Props) => {
+export const UserMetadata = ({ title, metaDataCount }: Props) => {
   // const showFollowers = () => {}
   // const showFollowing = () => {}
 
   return (
     <div className='cursor-pointer' onClick={() => {}}>
-      <Typography variant='bold_text_14'>{followers}</Typography>
+      <Typography variant='bold_text_14'>{metaDataCount}</Typography>
       <Typography variant='regular_text_14'>{title}</Typography>
     </div>
   )
